+++ src/xsdba/base.py
@@ -141,6 +141,12 @@ class Grouper(Parametrizable):
             by the `main_only` parameter of the `apply` method. If any of these dimensions are absent from the
             DataArrays, they will be omitted.
         """
+        if group == "time" and window > 1:
+            raise ValueError(
+                "The group given is 'time', but the window given is greater than 1. The `group = 'time'` option "
+                "takes the complete series, thus the concept of window is not applicable in this case. When using `group = 'time'`, "
+                "`window=1` is expected."
+            )
         if "." in group:
             dim, prop = group.split(".")
         else: